SQL improvement and some code refactoring

+++ sql/sqlite/ddl.sql
@@ -1,6 +1,6 @@
 ﻿BEGIN;
 CREATE TABLE person (
-  id        INTEGER PRIMARY KEY NOT NULL, -- 64 Bit
+  id        BIGINT PRIMARY KEY NOT NULL, -- 64 Bit
   created   TEXT, -- ISO8601 without Time Zone
   modified  TEXT, -- ISO8601 without Time Zone
   version   INTEGER NOT NULL DEFAULT 0,
@@ -11,8 +11,8 @@ CREATE INDEX idx_person_name ON person(name);
 CREATE INDEX idx_person_gender ON person(gender);
 
 CREATE TABLE person_parent (
-  person_id   INTEGER NOT NULL,
-  parent_id   INTEGER NOT NULL,
+  person_id   BIGINT NOT NULL,
+  parent_id   BIGINT NOT NULL,
   FOREIGN KEY (person_id) REFERENCES person(id),
   FOREIGN KEY (parent_id) REFERENCES person(id)
 );
@@ -21,12 +21,12 @@ CREATE INDEX idx_person_parent_person_id ON person_parent(person_id);
 CREATE INDEX idx_person_parent_parent_id ON person_parent(parent_id);
 
 CREATE TABLE person_photo (
-  id        INTEGER PRIMARY KEY NOT NULL, -- 64 Bit
+  id        BIGINT PRIMARY KEY NOT NULL, -- 64 Bit
   created   TEXT, -- ISO8601 without Time Zone
   modified  TEXT, -- ISO8601 without Time Zone
   version   INTEGER NOT NULL DEFAULT 0,
   data      BLOB,
-  person_id INTEGER,
+  person_id BIGINT,
   FOREIGN KEY (person_id) REFERENCES person(id)
 ) WITHOUT ROWID;
